+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## v1.?.? - 2024-??-?? - ???
 
 * Fix CAA rdata parsing to allow values with tags
+* Improve TXT (and SPF) record handling of unexpected whitespace
+  before/after/in-between quotes.
 
 ## v1.7.0 - 2024-04-29 - All the knobs and dials